# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WP Bridge AI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Conversion project management
  - name: parse
    description: Site parsing operations
  - name: wordpress
    description: WordPress sync operations
  - name: plugin
    description: WordPress plugin generation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ---- Admin portal -------------------------------------------------------
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Sign in to the admin portal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMeResponse"
        "401": { description: Invalid credentials }
        "403": { description: Account is not an admin }

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Sign out
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean } }
                required: [ok]

  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Probe the current admin session
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminMeResponse" }
        "401": { description: Not signed in }
        "403": { description: Signed in but not admin }

  /admin/ai-settings:
    get:
      operationId: getAdminAiSettings
      tags: [admin]
      summary: Read global AI settings (admin only)
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AiSettingsPublic" }
        "401": { description: Not signed in }
        "403": { description: Not admin }
    put:
      operationId: updateAdminAiSettings
      tags: [admin]
      summary: Update global AI settings (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AiSettingsUpdate" }
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AiSettingsPublic" }

  /admin/ai-settings/test-key:
    post:
      operationId: testAdminAiKey
      tags: [admin]
      summary: Probe the configured (or supplied) OpenAI API key
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                apiKey:
                  type: string
                  description: Optional ad-hoc key to test instead of the saved one.
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AiKeyTestResult" }

  /admin/projects/{id}/ai-status:
    get:
      operationId: getAdminProjectAiStatus
      tags: [admin]
      summary: Last-run + cache stats for a project (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectAiRunStatus" }

  /admin/projects/{id}/reanalyze:
    post:
      operationId: reanalyzeProject
      tags: [admin]
      summary: Invalidate AI cache and re-run engines on a project's source HTML
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Re-analysis completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  usedAi: { type: boolean }
                  pageCount: { type: integer }
                  aiStatus: { $ref: "#/components/schemas/ProjectAiRunStatus" }
                required: [ok, usedAi, pageCount, aiStatus]

  /projects/{id}/ai-status:
    get:
      operationId: getProjectAiStatus
      tags: [projects]
      summary: Public AI status for a project (read-only, no secrets)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectAiPublicStatus" }

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /projects/{id}/parse:
    post:
      operationId: parseProject
      tags: [parse]
      summary: Parse uploaded HTML into structured JSON
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseProjectBody"
      responses:
        "200":
          description: Parsed site structure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseResult"

  /projects/{id}/wordpress-config:
    put:
      operationId: updateWordPressConfig
      tags: [wordpress]
      summary: Save WordPress connection settings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WordPressConfig"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}/test-connection:
    post:
      operationId: testWordPressConnection
      tags: [wordpress]
      summary: Test WordPress REST API connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResult"

  /projects/{id}/push:
    post:
      operationId: pushToWordPress
      tags: [wordpress]
      summary: Push parsed site structure to WordPress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Push result with per-page outcomes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushResult"

  /projects/{id}/plugin:
    get:
      operationId: generatePlugin
      tags: [plugin]
      summary: Generate WordPress plugin PHP file for this project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plugin PHP content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginOutput"

  /projects/{id}/plugin-zip:
    get:
      operationId: downloadPluginZip
      tags: [plugin]
      summary: Download installable WordPress plugin ZIP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plugin ZIP ready for WordPress Upload Plugin
          content:
            application/zip:
              schema:
                type: string
                format: binary

  /projects/{id}/upload-zip:
    post:
      operationId: uploadProjectZip
      tags: [parse]
      summary: Upload a ZIP archive of an HTML site and parse its index page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Extracted and parsed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZipUploadResult"

  /projects/{id}/astro-export:
    get:
      operationId: exportProjectAstro
      tags: [parse]
      summary: Download the project as an Astro site ZIP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Astro project ZIP
          content:
            application/zip:
              schema:
                type: string
                format: binary

  /projects/stats:
    get:
      operationId: getProjectStats
      tags: [projects]
      summary: Aggregate stats across all projects
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"

components:
  schemas:
    AdminLoginRequest:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
      required: [username, password]

    AdminMeResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            id: { type: integer }
            username: { type: string }
            isAdmin: { type: boolean }
          required: [id, username, isAdmin]
      required: [user]

    AiSettingsPublic:
      type: object
      description: Safe-for-client view of global AI settings. The API key itself is never returned in plaintext.
      properties:
        enabled: { type: boolean }
        hasKey: { type: boolean }
        keyPreview:
          type: string
          nullable: true
          description: Display-only mask of the configured key — never the real key.
        model: { type: string }
        maxTokens: { type: integer }
        masterControllerMode: { type: boolean }
        status:
          type: string
          enum: [connected, invalid_key, disabled, unknown]
        statusMessage: { type: string, nullable: true }
        lastTestedAt: { type: string, nullable: true }
        updatedAt: { type: string }
      required: [enabled, hasKey, keyPreview, model, maxTokens, masterControllerMode, status, statusMessage, lastTestedAt, updatedAt]

    AiSettingsUpdate:
      type: object
      properties:
        enabled: { type: boolean }
        apiKey:
          type: string
          nullable: true
          description: Plaintext API key. Stored encrypted at rest. Pass null or empty string to clear.
        model: { type: string }
        maxTokens: { type: integer }
        masterControllerMode: { type: boolean }

    AiKeyTestResult:
      type: object
      properties:
        valid: { type: boolean }
        message: { type: string }
        settings: { $ref: "#/components/schemas/AiSettingsPublic" }
      required: [valid, message, settings]

    ProjectAiRunStatus:
      type: object
      properties:
        lastRunAt: { type: string, nullable: true }
        cacheEntries: { type: integer }
      required: [lastRunAt, cacheEntries]

    ProjectAiPublicStatus:
      type: object
      properties:
        aiEnabled: { type: boolean }
        aiStatus:
          type: string
          enum: [connected, invalid_key, disabled, unknown]
        model: { type: string }
        lastRunAt: { type: string, nullable: true }
        cacheEntries: { type: integer }
      required: [aiEnabled, aiStatus, model, lastRunAt, cacheEntries]

    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [created, parsed, configured, pushed, error]
        wpUrl:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
        pageCount:
          type: integer
          nullable: true
        lastPushedAt:
          type: string
          nullable: true
      required:
        - id
        - name
        - status
        - createdAt
        - updatedAt

    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            parsedSite:
              $ref: "#/components/schemas/ParsedSite"
              nullable: true
            designSystem:
              $ref: "#/components/schemas/DesignSystem"
              nullable: true
            wpConfig:
              $ref: "#/components/schemas/WordPressConfig"
              nullable: true
            pushLog:
              type: array
              items:
                $ref: "#/components/schemas/PushLogEntry"

    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name

    ParseProjectBody:
      type: object
      properties:
        htmlContent:
          type: string
          description: Raw HTML string to parse
        sourceType:
          type: string
          enum: [html, url]
          default: html
      required:
        - htmlContent

    ParseResult:
      type: object
      properties:
        parsedSite:
          $ref: "#/components/schemas/ParsedSite"
        designSystem:
          $ref: "#/components/schemas/DesignSystem"
        wpStructure:
          $ref: "#/components/schemas/WpStructure"
      required:
        - parsedSite
        - designSystem
        - wpStructure

    ParsedSite:
      type: object
      properties:
        pages:
          type: array
          items:
            $ref: "#/components/schemas/ParsedPage"
      required:
        - pages

    ParsedPage:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/ParsedSection"
      required:
        - name
        - slug
        - sections

    ParsedSection:
      type: object
      properties:
        type:
          type: string
          enum: [hero, about, features, testimonials, pricing, faq, cta, footer, gallery, blog, contact, custom]
        content:
          type: object
          additionalProperties: true
        rawHtml:
          type: string
      required:
        - type
        - content

    DesignSystem:
      type: object
      properties:
        font:
          type: string
        colors:
          type: array
          items:
            type: string
        buttonStyle:
          type: string
        headingStyle:
          type: string
      required:
        - font
        - colors
        - buttonStyle
        - headingStyle

    WpStructure:
      type: object
      properties:
        pages:
          type: array
          items:
            $ref: "#/components/schemas/WpPage"
      required:
        - pages

    WpPage:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/WpBlock"
      required:
        - title
        - slug
        - blocks

    WpBlock:
      type: object
      properties:
        blockType:
          type: string
        acfGroup:
          type: string
          nullable: true
        fields:
          type: object
          additionalProperties: true
        innerBlocks:
          type: array
          items:
            type: object
            additionalProperties: true
      required:
        - blockType
        - fields

    WordPressConfig:
      type: object
      properties:
        wpUrl:
          type: string
        wpUsername:
          type: string
          nullable: true
        wpAppPassword:
          type: string
          nullable: true
        wpApiKey:
          type: string
          nullable: true
        authMode:
          type: string
          enum: [basic, api_key]
          default: basic
        useAcf:
          type: boolean
          default: true
      required:
        - wpUrl

    ConnectionTestResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        wpVersion:
          type: string
          nullable: true
        siteTitle:
          type: string
          nullable: true
        pluginVersion:
          type: string
          nullable: true
          description: Companion plugin version reported by the WordPress install (api_key mode only).
        expectedPluginVersion:
          type: string
          nullable: true
          description: Plugin version this server expects.
        pluginOutdated:
          type: boolean
          nullable: true
          description: True when pluginVersion is older than expectedPluginVersion.
      required:
        - success
        - message

    PushResult:
      type: object
      properties:
        success:
          type: boolean
        pagesCreated:
          type: integer
        pagesUpdated:
          type: integer
        mediaUploaded:
          type: integer
        errors:
          type: array
          items:
            type: string
        wpPageUrls:
          type: array
          items:
            type: string
      required:
        - success
        - pagesCreated
        - pagesUpdated
        - mediaUploaded
        - errors
        - wpPageUrls

    PushLogEntry:
      type: object
      properties:
        pageName:
          type: string
        status:
          type: string
          enum: [success, error, skipped]
        wpId:
          type: integer
          nullable: true
        wpUrl:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - pageName
        - status
        - createdAt

    PluginOutput:
      type: object
      properties:
        phpCode:
          type: string
        filename:
          type: string
        apiKey:
          type: string
      required:
        - phpCode
        - filename
        - apiKey

    ZipUploadResult:
      type: object
      properties:
        fileCount:
          type: integer
        indexPath:
          type: string
        parsedSite:
          $ref: "#/components/schemas/ParsedSite"
        designSystem:
          $ref: "#/components/schemas/DesignSystem"
        wpStructure:
          $ref: "#/components/schemas/WpStructure"
      required:
        - fileCount
        - indexPath
        - parsedSite
        - designSystem
        - wpStructure

    ProjectStats:
      type: object
      properties:
        totalProjects:
          type: integer
        parsedProjects:
          type: integer
        pushedProjects:
          type: integer
        totalPagesConverted:
          type: integer
      required:
        - totalProjects
        - parsedProjects
        - pushedProjects
        - totalPagesConverted