# lib/api-zod/src/generated/types/aiKeyTestResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WP Bridge AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiSettingsPublic } from "./aiSettingsPublic";

export interface AiKeyTestResult {
  valid: boolean;
  message: string;
  settings: AiSettingsPublic;
}