# lib/api-zod/src/generated/types/aiSettingsUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WP Bridge AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiSettingsUpdate {
  enabled?: boolean;
  /** Plaintext API key. Stored encrypted at rest. Pass null or empty string to clear. */
  apiKey?: string | null;
  model?: string;
  maxTokens?: number;
  masterControllerMode?: boolean;
}