# lib/api-zod/src/generated/types/projectAiPublicStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WP Bridge AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectAiPublicStatusAiStatus } from "./projectAiPublicStatusAiStatus";

export interface ProjectAiPublicStatus {
  aiEnabled: boolean;
  aiStatus: ProjectAiPublicStatusAiStatus;
  model: string;
  lastRunAt: string | null;
  cacheEntries: number;
}