# lib/api-zod/src/generated/types/projectAiRunStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WP Bridge AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectAiRunStatus {
  lastRunAt: string | null;
  cacheEntries: number;
}